aws_resources/requirements.txt
aws-cdk.core
aws_cdk.aws_sns
aws_cdk.aws_sns_subscriptions
aws_cdk.aws_lambda
aws_cdk.aws_logs
aws_cdk.aws_sqs